Set up get and post methods for user database. Save user currently not liking json format

+++ backEnd/src/main/java/com/wellnessapp/demo/UserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
 
-public interface UserRepository extends MongoRepository<User, Long> {
+public interface UserRepository extends MongoRepository<User, Integer> {
     // possible functions for querrying the db
     public User findByUsername(String username);
     public List<User> findByState(String State);
